refactor: 뷰포트 높이 로직 -> 커스텀 훅으로 구현
