+++ src/App.tsx
@@ -3,26 +3,15 @@ import Login from '@/pages/Login'
 import AuthCallback from '@/pages/AuthCallback'
 import Nickname from './pages/Nickname'
 import NicknameComplete from './pages/NicknameComplete'
-import { useEffect } from 'react'
+import Main from './pages/Main'
+import useViewportHeight from './hooks/useViewportHeight'
 
 export default function App() {
-  useEffect(() => {
-    const setVh = () => {
-      const vh = window.innerHeight * 0.01
-      document.documentElement.style.setProperty('--vh', `${vh}px`)
-    }
-
-    setVh()
-    window.addEventListener('resize', setVh)
-
-    return () => {
-      window.removeEventListener('resize', setVh)
-    }
-  }, [])
+  useViewportHeight()
 
   return (
     <Routes>
-      <Route index element={<div>홈</div>} />
+      <Route index element={<Main />} />
       <Route path="login" element={<Login />} />
       <Route path="callback/kakaotalk" element={<AuthCallback />} />
       <Route path="nickname" element={<Nickname />} />
