+++ src/hooks/useViewportHeight.ts
@@ -1,0 +1,19 @@
+import { useEffect } from 'react'
+
+function useViewportHeight() {
+  useEffect(() => {
+    const setVh = () => {
+      const vh = window.innerHeight * 0.01
+      document.documentElement.style.setProperty('--vh', `${vh}px`)
+    }
+
+    setVh()
+    window.addEventListener('resize', setVh)
+
+    return () => {
+      window.removeEventListener('resize', setVh)
+    }
+  }, [])
+}
+
+export default useViewportHeight
